+++ packages/plugins/@nocobase/plugin-calendar/src/client/schema-initializer/CalendarBlockProvider.tsx
@@ -1,5 +1,5 @@
 import { ArrayField } from '@formily/core';
-import { useField } from '@formily/react';
+import { useField, useFieldSchema } from '@formily/react';
 import { BlockProvider, FixedBlockWrapper, useBlockRequestContext, useParsedFilter } from '@nocobase/client';
 import _ from 'lodash';
 import React, { createContext, useContext, useEffect, useMemo } from 'react';
@@ -39,8 +39,20 @@ const InternalCalendarBlockProvider = (props) => {
   );
 };
 
+const useCompatCalendarBlockParams = (props) => {
+  const fieldSchema = useFieldSchema();
+
+  // 因为 x-use-decorator-props 的值是固定不变的，所以可以在条件中使用 hooks
+  if (fieldSchema['x-use-decorator-props']) {
+    return props.params;
+  } else {
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    return useCalendarBlockParams(props);
+  }
+};
+
 export const CalendarBlockProvider = (props) => {
-  const params = useCalendarBlockParams(props);
+  const params = useCompatCalendarBlockParams(props);
   return (
     <BlockProvider name="calendar" {...props} params={params}>
       <InternalCalendarBlockProvider {...props} />
