+++ packages/plugins/@nocobase/plugin-kanban/src/client/__tests__/createKanbanBlockSchema.test.ts
@@ -1,4 +1,4 @@
-import { createKanbanBlockSchema } from '../utils';
+import { createKanbanBlockSchema } from '../createKanbanBlockSchema';
 
 vi.mock('@formily/shared', () => {
   return {
